Updatded after merge with 1726

+++ src/main/java/cc/altius/FASP/model/rowMapper/NewSupplyPlanBatchResultSetExtractor.java
@@ -56,7 +56,7 @@ public class NewSupplyPlanBatchResultSetExtractor implements ResultSetExtractor<
             if (rs.wasNull()) {
                 bd.setStock(null);
             }
-            bd.setInventoryQty(rs.getLong("INVENTORY_QTY"));
+            bd.setInventoryQty(rs.getDouble("INVENTORY_QTY"));
             if (rs.wasNull()) {
                 bd.setInventoryQty(null);
             }
